Move submenus to discrete imports. Add Call vote menu.
