+++ frontend/src/component/menu/SteamIDMenu.tsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import SteamID from 'steamid';
+import { useTranslation } from 'react-i18next';
+import { IconMenuItem, NestedMenuItem } from 'mui-nested-menu';
+import ArrowRightOutlinedIcon from '@mui/icons-material/ArrowRightOutlined';
+import ContentCopyOutlinedIcon from '@mui/icons-material/ContentCopyOutlined';
+import FlagIcon from '@mui/icons-material/Flag';
+import { logError, writeToClipboard } from '../../util';
+import { SteamIDProps, SubMenuProps } from './common';
+
+export const SteamIDMenu = ({
+    contextMenuPos,
+    steam_id,
+    onClose
+}: SubMenuProps & SteamIDProps) => {
+    const { t } = useTranslation();
+    const id = new SteamID(steam_id);
+    return (
+        <NestedMenuItem
+            rightIcon={<ArrowRightOutlinedIcon />}
+            leftIcon={<ContentCopyOutlinedIcon color={'primary'} />}
+            label={t('player_table.menu.copy_label')}
+            parentMenuOpen={contextMenuPos !== null}
+        >
+            {[
+                id.getSteam2RenderedID(),
+                id.getSteam3RenderedID(),
+                id.getSteamID64()
+            ].map((sid) => {
+                return (
+                    <IconMenuItem
+                        leftIcon={<FlagIcon color={'primary'} />}
+                        onClick={async () => {
+                            try {
+                                await writeToClipboard(sid);
+                            } catch (e) {
+                                logError(e);
+                            } finally {
+                                onClose();
+                            }
+                        }}
+                        label={sid}
+                    />
+                );
+            })}
+        </NestedMenuItem>
+    );
+};
