+++ frontend/src/component/menu/common.ts
@@ -1,0 +1,13 @@
+export interface SubMenuProps {
+    contextMenuPos: NullablePosition;
+    onClose: () => void;
+}
+
+export interface SteamIDProps {
+    steam_id: string;
+}
+
+export type NullablePosition = {
+    mouseX: number;
+    mouseY: number;
+} | null;
